fix: every loop index define on size_t

+++ ft_memcpy.c
@@ -6,7 +6,7 @@
 /*   By: dcastor <dcastor@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/23 17:34:48 by dcastor           #+#    #+#             */
-/*   Updated: 2025/04/23 17:41:46 by dcastor          ###   ########.fr       */
+/*   Updated: 2025/04/24 10:22:45 by dcastor          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memcpy(void *dest, const void *src, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < n)
